Replace "product" with "service-type"
This brings the cli up to date with the changes from the last week in
the API where products are now referred to as service types.

DBAAS-8407
