+++ cmd/constants.go
@@ -34,8 +34,8 @@ const (
 	SERVICES        = "services"
 	IP_ADDRESS      = "ip-address"
 	MACHINE         = "machine"
-	PRODUCT         = "product"
-	PRODUCTS        = "products"
+	SERVICE_TYPE    = "service-type"
+	SERVICE_TYPES   = "service-types"
 	PROVIDER        = "provider"
 	PROVIDERS       = "providers"
 	QUOTA           = "quota"
